Switches to unicode, adds a .gitignore for VS
The old version of the program took input and gave output in ASCII.
This commit replaces this with unicode.
The purpose of this is opening up for the usage of special characters.
The commit also adds a .gitignore in the csharp-directory.
The .gitignore is the  recommendations for Microsoft Visual Studio.